ANPR_with_opencv-main/camera_client/requirements.txt
opencv-python
requests
pyserial
pillow
easyocr
